store decorators
did some thinking on how ‘stateful actions’ for stores, wrote a
treatment for it, then scrapped it.

thinking of going with decorators, but i’m unsure of the convention I’d
like to have; is any function on a store an action? should only
functions marked @action be allowed to set state? or should there just
be a @stateful decorator for stateful actions? Dunno, need to think.